Revert "Reorganize pp"
