+++ dynamo/preprocessing/Preprocessor.py
@@ -25,19 +25,21 @@ from .external import (
     select_genes_by_pearson_residuals,
 )
 from .gene_selection import select_genes_by_seurat_recipe, select_genes_monocle
-from .normalization import calc_sz_factor, normalize
-from .pca import pca
-from .QC import (
-    filter_cells_by_outliers as monocle_filter_cells_by_outliers,
-    regress_out_parallel,
+from .preprocess import normalize_cell_expr_by_size_factors_legacy, pca
+from .preprocessor_utils import (
+    Freeman_Tukey,
+    _infer_labeling_experiment_type,
+    calc_sz_factor,
 )
-from .QC import (
+from .preprocessor_utils import (
+    filter_cells_by_outliers as monocle_filter_cells_by_outliers,
+)
+from .preprocessor_utils import (
     filter_genes_by_outliers as monocle_filter_genes_by_outliers,
 )
-from .QC import basic_stats
-from .transform import Freeman_Tukey, log, log1p, log2
+from .preprocessor_utils import log, log1p, log2, normalize, regress_out_parallel
 from .utils import (
-    _infer_labeling_experiment_type,
+    basic_stats,
     collapse_species_adata,
     convert2symbol,
     convert_layers2csr,
