+++ dynamo/preprocessing/__init__.py
@@ -10,31 +10,30 @@ from .external import (
     sctransform,
     select_genes_by_pearson_residuals,
 )
-from .normalization import normalize
-from .QC import (
-    basic_stats,
-    filter_genes_by_clusters,
-    filter_cells_by_outliers,
-    filter_genes_by_outliers,
-    filter_genes_by_pattern,
-)
-from .pca import pca, top_pca_genes
-from .transform import log1p, log1p_adata_layer
-from .utils import (
-    compute_gene_exp_fraction,
-    convert2symbol,
-    decode,
-    get_svr_filter,
-    relative2abs,
-    scale,
-)
-from .deprecated import (
-    cook_dist,
+from .preprocess import (
     calc_sz_factor_legacy,
-    normalize_cell_expr_by_size_factors_legacy,
+    filter_cells_by_outliers,
     filter_cells_legacy,
+    filter_genes_by_clusters_,
+    filter_genes_by_outliers,
+    get_svr_filter,
+    highest_frac_genes,
+    normalize_cell_expr_by_size_factors_legacy,
     recipe_monocle,
     recipe_velocyto,
+)
+from .preprocessor_utils import *
+from .utils import (
+    basic_stats,
+    compute_gene_exp_fraction,
+    convert2symbol,
+    cook_dist,
+    decode,
+    filter_genes_by_pattern,
+    pca,
+    relative2abs,
+    scale,
+    top_pca_genes,
 )
 
 filter_cells = filter_cells_by_outliers
@@ -43,7 +42,7 @@ log1p = log1p
 normalize_cells = normalize
 
 from .CnmfPreprocessor import CnmfPreprocessor
-from .gene_selection import calc_Gini, calc_dispersion_by_svr, highest_frac_genes, select_genes_monocle
+from .gene_selection import calc_Gini, calc_dispersion_by_svr, select_genes_monocle
 from .Preprocessor import Preprocessor
 
 __all__ = [
@@ -63,7 +62,7 @@ __all__ = [
     "select_genes_by_pearson_residuals",
     "filter_genes",
     "filter_genes_by_outliers",
-    "filter_genes_by_clusters",
+    "filter_genes_by_clusters_",
     "calc_dispersion_by_svr",
     "get_svr_filter",
     "highest_frac_genes",
